# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Intelligent scheduling system for UK personal trainers.",
        "title": "PT Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PT Scheduler Support",
            "email": "support@pt-scheduler.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clients/{clientID}/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a client's profile including their sessions-per-month quota and priority score.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Get client profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client UUID",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_users.ClientProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a client's full name, phone, and timezone.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Update client profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client UUID",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_users.UpdateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_users.ClientProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/coaches/{coachID}/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the public profile of a coach, including business name and user details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coaches"
                ],
                "summary": "Get coach profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coach UUID",
                        "name": "coachID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_users.CoachProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the coach's business name, full name, phone, and timezone.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coaches"
                ],
                "summary": "Update coach profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coach UUID",
                        "name": "coachID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_users.UpdateCoachRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_users.CoachProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all data held for the currently authenticated user as a JSON document (GDPR right of access).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gdpr"
                ],
                "summary": "Export my data (GDPR)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_users.DataExport"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_users.Client": {
            "type": "object",
            "properties": {
                "coach_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "priority_score": {
                    "type": "integer"
                },
                "sessions_per_month": {
                    "type": "integer"
                },
                "tenure_started_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_users.ClientProfile": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/internal_users.Client"
                },
                "user": {
                    "$ref": "#/definitions/internal_users.User"
                }
            }
        },
        "internal_users.Coach": {
            "type": "object",
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_users.CoachProfile": {
            "type": "object",
            "properties": {
                "coach": {
                    "$ref": "#/definitions/internal_users.Coach"
                },
                "user": {
                    "$ref": "#/definitions/internal_users.User"
                }
            }
        },
        "internal_users.DataExport": {
            "type": "object",
            "properties": {
                "exported_at": {
                    "type": "string"
                },
                "payments": {},
                "profile": {},
                "sessions": {},
                "user": {}
            }
        },
        "internal_users.UpdateClientRequest": {
            "type": "object",
            "required": [
                "full_name",
                "timezone"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "internal_users.UpdateCoachRequest": {
            "type": "object",
            "required": [
                "full_name",
                "timezone"
            ],
            "properties": {
                "business_name": {
                    "type": "string",
                    "maxLength": 120
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "internal_users.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer \" followed by your access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}